First version of types
